requirements.txt
APScheduler==3.6.3
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
confluent_kafka==1.4.2
cycler==0.10.0
datetime==4.3
distlib==0.3.1
filelock==3.0.12
flask==1.1.2
fleep==1.0.1
future==0.18.2
html5lib==1.1
idna==2.10
imgur-python==0.2.1
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
kiwisolver==1.2.0
line-bot-sdk==1.16.0
lxml==4.5.1
matplotlib==3.2.2
mpl-finance==0.10.1
mplfinance==0.12.6a3
multitasking==0.0.9
numpy==1.19.0
pandas==1.0.5
pandas-datareader==0.8.1
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
selenium==3.141.0
toml==0.10.1
tox==3.16.1
tox-travis==0.12
twder==0.1.4
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.25
webencodings==0.5.1
yfinance==0.1.54
zipp==3.1.0
zope.interface==5.1.0












